+++ src/migrations/m180423_175007_craft2.php
@@ -1,0 +1,155 @@
+<?php
+namespace fruitstudios\linkit\migrations;
+
+use fruitstudios\linkit\Linkit;
+use fruitstudios\linkit\fields\LinkitField;
+use fruitstudios\linkit\models\Email;
+use fruitstudios\linkit\models\Phone;
+use fruitstudios\linkit\models\Url;
+use fruitstudios\linkit\models\Entry;
+use fruitstudios\linkit\models\Category;
+use fruitstudios\linkit\models\Asset;
+use fruitstudios\linkit\models\Product;
+
+use Craft;
+use craft\db\Migration;
+use craft\helpers\Json;
+
+class m180423_175007_craft2 extends Migration
+{
+
+    public function safeUp()
+    {
+        if ($this->_upgradeFromCraft2()) {
+            return;
+        }
+    }
+
+    private function _upgradeFromCraft2()
+    {
+        // Locate and remove old linkit
+        $row = (new \craft\db\Query())
+            ->select(['id', 'settings'])
+            ->from(['{{%plugins}}'])
+            ->where(['in', 'handle', ['fruitlinkit', 'fruit-link-it', 'fruit-linkit']])
+            ->one();
+
+        if($row)
+        {
+            $this->delete('{{%plugins}}', ['id' => $row['id']]);
+        }
+
+        // Look for any old linkit fields and update their settings
+        $fields = (new \craft\db\Query())
+            ->select(['id', 'settings'])
+            ->from(['{{%fields}}'])
+            ->where(['in', 'type', ['FruitLinkIt']])
+            ->all();
+
+        if($fields)
+        {
+            // Update field settings
+            foreach($fields as $field)
+            {
+                $oldSettings = $field['settings'] ? Json::decode($field['settings']) : null;
+                $newSettings = $this->_migrateFieldSettings($oldSettings);
+
+                $this->update('{{%fields}}', [
+                    'type' => LinkitField::class,
+                    'settings' => Json::encode($newSettings)
+                ], ['id' => $field['id']]);
+            }
+        }
+
+        return true;
+    }
+
+    private function _migrateFieldSettings($oldSettings)
+    {
+        if(!$oldSettings)
+        {
+            return null;
+        }
+
+        $linkitField = new LinkitField();
+
+        $newSettings = $linkitField->getSettings();
+        $newSettings['defaultText'] = $oldSettings['defaultText'] ?? '';
+        $newSettings['allowTarget'] = $oldSettings['allowTarget'] ?? 0;
+        $newSettings['allowCustomText'] = $oldSettings['allowCustomText'] ?? 0;
+
+        if($oldSettings['types'])
+        {
+            foreach ($oldSettings['types'] as $oldType)
+            {
+                switch ($oldType)
+                {
+                    case 'email':
+                        $newSettings['types'][Email::class] = [
+                            'enabled' => 1,
+                            'customLabel' => null,
+                        ];
+                        break;
+
+                    case 'custom':
+                        $newSettings['types'][Url::class] = [
+                            'enabled' => 1,
+                            'customLabel' => null,
+                        ];
+                        break;
+
+                    case 'tel':
+                        $newSettings['types'][Phone::class] = [
+                            'enabled' => 1,
+                            'customLabel' => null,
+                        ];
+                        break;
+
+                    case 'entry':
+                        $newSettings['types'][Entry::class] = [
+                            'enabled' => 1,
+                            'customLabel' => null,
+                            'sources' => $oldSettings['entrySources'] ?? '*',
+                            'customSelectionLabel' => $oldSettings['entrySelectionLabel'] ?? '',
+                        ];
+                        break;
+
+                    case 'category':
+                        $newSettings['types'][Category::class] = [
+                            'enabled' => 1,
+                            'customLabel' => null,
+                            'sources' => $oldSettings['categorySources'] ?? '*',
+                            'customSelectionLabel' => $oldSettings['categorySelectionLabel'] ?? '',
+                        ];
+                        break;
+
+                    case 'asset':
+                        $newSettings['types'][Asset::class] = [
+                            'enabled' => 1,
+                            'customLabel' => null,
+                            'sources' => $oldSettings['assetSources'] ?? '*',
+                            'customSelectionLabel' => $oldSettings['assetSelectionLabel'] ?? '',
+                        ];
+                        break;
+
+                    case 'product':
+                        $newSettings['types'][Product::class] = [
+                            'enabled' => 1,
+                            'customLabel' => null,
+                            'sources' => $oldSettings['entrySources'] ?? '*',
+                            'customSelectionLabel' => $oldSettings['entrySelectionLabel'] ?? '',
+                        ];
+                        break;
+                }
+            }
+        }
+
+        return $newSettings;
+    }
+
+    public function safeDown()
+    {
+        echo "m180423_175007_craft2 cannot be reverted.\n";
+        return false;
+    }
+}